Update COM and UFC library

+++ UFC/MDS/MColumn.cpp
@@ -199,7 +199,7 @@ const double MColumn::ToDouble( void )
 	}
 	else if ( FDefinition->FType == ctInt64 ) ///< ctInt64
 	{
-		Int64 int64Value;
+		int int64Value;
 		memcpy( &int64Value, FData, sizeof(Int64) );
 		Result = (double)int64Value;
 	}
